# backend/openapi.yaml
openapi: 3.0.3
info:
  title: CivicPulse API
  version: 1.0.0
  description: |
    API contract between the Flask backend, the React frontend and the AI worker.
    Conventions:
    - All paths are under /api/v1. JSON in and out, except POST /reports (multipart).
    - Auth uses a JWT [a signed token the app sends with each request to prove who the user is]
      in the header `Authorization: Bearer <token>`.
    - Times are ISO 8601 UTC strings. Locations are decimal degrees (lat, lon).
    - Lists are paginated with `page` and `per_page` and return {items, page, per_page, total}.
    - Errors always look like {error: {code, message, details}}.
    - Status flow: reported -> verified -> assigned -> in_progress -> resolved.
      A verifier may also reject a reported item (status rejected).
servers:
  - url: http://localhost:5000/api/v1
    description: Local development
...
security:
  - bearerAuth: []
paths:
...
  /reports/{id}/history:
    parameters:
      - {$ref: '#/components/parameters/ReportId'}
    get:
      tags: [Workflow]
      summary: Status trail of a report
      responses:
        '200':
          description: Every status change, oldest first
          content:
            application/json:
              schema:
                type: array
                items: {$ref: '#/components/schemas/StatusEvent'}
        '404': {$ref: '#/components/responses/NotFound'}
  /reports/{id}/status:
    parameters:
      - {$ref: '#/components/parameters/ReportId'}
    patch:
      tags: [Workflow]
      summary: Change status (verifier, authority, admin)
      description: |
        Allowed moves: reported->verified, reported->rejected, assigned->in_progress, in_progress->resolved.
        Moving to assigned is done only through POST /reports/{id}/assign (409 use_assign_endpoint).
        Verifier or admin: verified and rejected (rejecting needs a note). Assigned authority or admin:
        in_progress and resolved. rejected and resolved are final. A merged duplicate cannot be updated
        directly (409 is_duplicate). Each change creates a status event and notifies followers.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [status]
              properties:
                status: {$ref: '#/components/schemas/Status'}
                note: {type: string, maxLength: 500}
      responses:
        '200':
          description: Updated report
          content:
            application/json:
              schema: {$ref: '#/components/schemas/Report'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '404': {$ref: '#/components/responses/NotFound'}
        '409': {$ref: '#/components/responses/Conflict'}
        '422': {$ref: '#/components/responses/Invalid'}
  /reports/{id}/assign:
    parameters:
      - {$ref: '#/components/parameters/ReportId'}
    post:
      tags: [Workflow]
      summary: Assign to an authority user (verifier, authority, admin)
      description: |
        The report must be verified. assignee_id must be a user with the authority role (422 otherwise).
        due_date is optional and cannot be in the past. A report that is assigned but not started can be
        reassigned with the same call; once in_progress it cannot (409 cannot_reassign).
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [assignee_id]
              properties:
                assignee_id: {type: integer}
                due_date: {type: string, format: date}
      responses:
        '200':
          description: Updated report (status becomes assigned)
          content:
            application/json:
              schema: {$ref: '#/components/schemas/Report'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '409': {$ref: '#/components/responses/Conflict'}
...
components:
  securitySchemes:
    bearerAuth: {type: http, scheme: bearer, bearerFormat: JWT}
  parameters:
    ReportId:
      name: id
      in: path
      required: true
      schema: {type: integer}
  responses:
...
    Forbidden:
      description: Your role cannot do this
      content: {application/json: {schema: {$ref: '#/components/schemas/Error'}}}
    NotFound:
      description: Not found
      content: {application/json: {schema: {$ref: '#/components/schemas/Error'}}}
    Conflict:
      description: Not allowed in the current state (for example an invalid status move)
      content: {application/json: {schema: {$ref: '#/components/schemas/Error'}}}
    Invalid:
      description: Validation failed
      content: {application/json: {schema: {$ref: '#/components/schemas/Error'}}}
  schemas:
    Error:
      type: object
      properties:
        error:
          type: object
          properties:
            code: {type: string, example: invalid_transition}
            message: {type: string}
            details: {type: object}
...
    Status: {type: string, enum: [reported, verified, rejected, assigned, in_progress, resolved]}
...
    Category:
      type: object
      properties:
        id: {type: integer}
        name: {type: string, example: Pothole}
        icon: {type: string, example: pothole}
...
    Location:
      type: object
      properties:
        lat: {type: number}
        lon: {type: number}
    Report:
      type: object
      properties:
        id: {type: integer}
        category: {$ref: '#/components/schemas/Category'}
        description: {type: string}
        status: {$ref: '#/components/schemas/Status'}
        location: {$ref: '#/components/schemas/Location'}
        area_id: {type: integer, nullable: true}
        photo_url: {type: string}
        created_by: {type: integer}
        assigned_to: {type: integer, nullable: true}
        due_date: {type: string, format: date, nullable: true}
        report_count: {type: integer, description: 'Residents who reported this same case (1 plus merged duplicates)'}
        duplicate_of: {type: integer, nullable: true, description: 'Parent report id if this was merged'}
        ai_suggested_category:
          type: object
          nullable: true
          properties:
            category_id: {type: integer}
            confidence: {type: number, minimum: 0, maximum: 1}
        is_following: {type: boolean}
        created_at: {type: string, format: date-time}
        updated_at: {type: string, format: date-time}
...
    StatusEvent:
      type: object
      properties:
        id: {type: integer}
        report_id: {type: integer}
        status: {$ref: '#/components/schemas/Status'}
        changed_by: {type: integer}
        note: {type: string, nullable: true}
        created_at: {type: string, format: date-time}
...